Neukolln loading complete
Still to do: authoritarianism and criminal history. Still, so many changes it deserves to go into master. The R procedure needs some cleanup but resisted my checks.

Now all the data is loaded not by the old file but by the new; probabilities are assigned with respect to the neighborhood population.