Loeading code from combination data
This is a prelimiarry version.
load-opinions is not yet called in the setup. We need a threahold to transform fundamntalism in authoritarian?

It assumes that the opinions file has the keys in the same order as in the netlogo code, then the three opnion scores, and finally the fundamentalism value.